Started tokenizer
