+++ Makefile
@@ -1,5 +1,5 @@
 dgnasm: dgnasm.h *.c
-	gcc -g -rdynamic -o dgnasm *.c
+	gcc -g -rdynamic -o dgnasm dgnasm.c
 
 clean:
 	rm -f *.o dgnasm
